+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added the `NetworkStatusToast` component to the store.
 
 ## [2.7.1] - 2019-03-08
 ### Removed
@@ -29,7 +31,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add updateOrderFormCheckin mutation to OrderFormProvider
 
 ## [2.5.6] - 2019-02-26
-
 ### Removed
 
 - Removed address-locator dependency.